Add message warning component
